pyproject.toml
[tool.poetry]
name = "mastodon-wordcloud"
version = "0.1.0"
description = ""
authors = ["kelvin27315"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/kelvin27315/mastodon-wordcloud"

[tool.poetry.dependencies]
python = "^3.11.4"
pandas = "^2.0.3"
mecab-python3 = "^1.0.6"
mastodon-py = "^1.8.1"
tqdm = "^4.65.1"
wordcloud = "^1.9.2"
japanize-matplotlib = "^1.1.3"
scikit-learn = "^1.3.0"
gensim = "^4.3.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
